updated the print_digit data type to int

+++ main.h
@@ -6,7 +6,7 @@
 
 /*function prototypes*/
 
-int print_digit(long long num);
+int print_digit(int num);
 int print_str(char *str);
 int put_s(char *str);
 int _printf(const char *format, ...);
